requirements_eda_annotate.txt
torch>=2.1.0
torchvision>=0.16.0
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
matplotlib>=3.7.0
tqdm>=4.65.0
hydra-core>=1.3.2
jupyterlab>=4.0.0
ipykernel>=6.29.0
ipywidgets>=8.0.0
ultralytics>=8.0.196
ensemble-boxes>=1.0.9
opencv-python>=4.8.0.74
wandb>=0.16.0
openpyxl>=3.1.2
supervision==0.26.1
transformers==4.57.1
addict==2.4.0
pycocotools==2.0.10
timm==1.0.21
yapf==0.43.0
